Registracija kroz validaciju email-a gotova, popravljeno brisanje zadataka

+++ app/Http/Controllers/TasksController.php
@@ -4,8 +4,11 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Task;
+use App\Models\UserTasks;
+use App\Models\UserAnsweredTasks;
 use App\Models\LevelLevel;
 use Illuminate\Support\MessageBag;
+use Auth;
 
 class TasksController extends Controller
 {
@@ -191,8 +194,21 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        $user_task = UserTasks::where('id_task', $id);
+        $user_answered_task = UserAnsweredTasks::where('id_answered_task', $id);
 
+        if($user_answered_task != null){
+            $user_answered_task->delete();
+        }
+
+        if($user_task != null){
+            $user_task->delete();
+        }
+
+        if($task != null){
+            $task->delete();
+        }
+        
         return redirect('/tasks/index/' . session('id_level'));
     }
 }
